Revert "Merging in Theme changes and moving win32Tests to Microsoft.Plugin.Program.UnitTests"
This reverts commit 030dfc2370f7d6df01a78ff3a00a54dd80bff0b8.
