--- src/modules/launcher/Plugins/Microsoft.Plugin.Program.UnitTests/Programs/PackageCatalogWrapperTests.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Microsoft.Plugin.Program.UnitTests.Programs
-{
-    class PackageCatalogWrapperTests
-    {
-    }
-}
