+++ src/modules/launcher/Plugins/Microsoft.Plugin.Program/Views/ProgramSetting.xaml.cs
@@ -51,6 +51,7 @@ namespace Microsoft.Plugin.Program.Views
             Task.Run(() =>
             {
                 Dispatcher.Invoke(() => { indexingPanel.Visibility = Visibility.Visible; });
+                Main.IndexPrograms();
                 Dispatcher.Invoke(() => { indexingPanel.Visibility = Visibility.Hidden; });
             });
         }
